+++ packages/plugin-ext/src/common/index.ts
@@ -14,10 +14,6 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-// Exports contribution point for uri postprocessor of hosted plugin manager.
-// This could be used to alter hosted instance uri, for example, change port.
-export * from '../hosted/node/hosted-plugin-uri-postprocessor';
-
 // Here we expose types from @theia/plugin, so it becames a direct dependency
 export * from '../common/plugin-protocol';
 export * from '../plugin/plugin-context';